+++ lib/algorithm/visitor/partition_creator_visitor.dart
@@ -1,8 +1,8 @@
-import 'package:push_puzzle/utility/bsp/area.dart';
-import 'package:push_puzzle/utility/bsp/dungeon_config.dart';
-import 'package:push_puzzle/utility/bsp/partition/partition.dart';
-import 'package:push_puzzle/utility/bsp/visitor/visitor.dart';
-import 'package:push_puzzle/utility/bsp/extention/list2d_extention.dart';
+import 'package:push_puzzle/algorithm/area.dart';
+import 'package:push_puzzle/algorithm/dungeon_config.dart';
+import 'package:push_puzzle/algorithm/partition/partition.dart';
+import 'package:push_puzzle/algorithm/visitor/visitor.dart';
+import 'package:push_puzzle/algorithm/extention/list2d_extention.dart';
 
 class PartitionCreatorVisitor extends Visitor {
   late DungeonConfig config = DungeonConfig();
